split zm and kzg rm old kzg
