+++ jolt-core/src/poly/commitment/kzg.rs
@@ -1,59 +1,33 @@
-use ark_ec::scalar_mul::fixed_base::FixedBase;
-use std::{borrow::Borrow, marker::PhantomData};
-
 use crate::msm::VariableBaseMSM;
-use crate::poly;
-use crate::poly::unipoly::UniPoly;
-use ark_ec::{pairing::Pairing, AffineRepr, CurveGroup};
+use crate::poly::{unipoly::UniPoly, field::JoltField};
+use crate::utils::errors::ProofVerifyError;
+use ark_ec::scalar_mul::fixed_base::FixedBase;
+use ark_ec::{pairing::Pairing, CurveGroup};
 use ark_ff::PrimeField;
-use ark_std::One;
 use ark_std::UniformRand;
-use rand_chacha::rand_core::RngCore;
-use thiserror::Error;
+use rand_core::{CryptoRng, RngCore};
+use std::marker::PhantomData;
+use std::sync::Arc;
 
-#[derive(Error, Debug)]
-pub enum KZGError {
-    #[error("Length Error: SRS Length: {0}, Key Length: {0}")]
-    KeyLengthError(usize, usize),
-    #[error("Length Error: Commitment Key Length: {0}, Polynomial Degree {0}")]
-    CommitLengthError(usize, usize),
-    #[error("Failed to compute quotient polynomial due to polynomial division")]
-    PolyDivisionError,
-}
+#[cfg(feature = "ark-msm")]
+use ark_ec::VariableBaseMSM;
 
-#[derive(Debug, Clone, Default)]
-pub struct UniversalKzgSrs<P: Pairing> {
+#[derive(Clone, Debug)]
+pub struct SRS<P: Pairing> {
     pub g1_powers: Vec<P::G1Affine>,
     pub g2_powers: Vec<P::G2Affine>,
 }
 
-#[derive(Clone, Debug)]
-pub struct KZGProverKey<P: Pairing> {
-    /// generators
-    pub g1_powers: Vec<P::G1Affine>,
-}
+impl<P: Pairing> SRS<P> {
+    pub fn setup<R: RngCore + CryptoRng>(mut rng: &mut R, max_degree: usize) -> Self {
+        let beta = P::ScalarField::rand(&mut rng);
+        let g1 = P::G1::rand(&mut rng);
+        let g2 = P::G2::rand(&mut rng);
 
-#[derive(Clone, Debug)]
-pub struct KZGCommitment<P: Pairing>(P::G1Affine);
-pub struct KZGVerifierKey<P: Pairing> {
-    /// The generator of G1.
-    pub g1: P::G1Affine,
-    /// The generator of G2.
-    pub g2: P::G2Affine,
-    /// tau times the above generator of G2.
-    pub tau_2: P::G2Affine,
-}
-
-impl<P: Pairing> UniversalKzgSrs<P> {
-    pub fn setup<R: RngCore>(max_degree: usize, rng: &mut R) -> UniversalKzgSrs<P> {
-        let tau = P::ScalarField::rand(rng);
-        let g1 = P::G1::rand(rng);
-        let g2 = P::G2::rand(rng);
-
-        let tau_powers: Vec<_> = (0..=max_degree)
-            .scan(tau, |state, _| {
-                let val = *state;
-                *state *= &tau;
+        let beta_powers: Vec<P::ScalarField> = (0..=max_degree)
+            .scan(beta, |acc, _| {
+                let val = *acc;
+                *acc *= beta;
                 Some(val)
             })
             .collect();
@@ -61,174 +35,186 @@ impl<P: Pairing> UniversalKzgSrs<P> {
         let window_size = FixedBase::get_mul_window_size(max_degree);
         let scalar_bits = P::ScalarField::MODULUS_BIT_SIZE as usize;
 
-        //TODO: gate with rayon
-        let g1_table = FixedBase::get_window_table(scalar_bits, window_size, g1);
-        let g2_table = FixedBase::get_window_table(scalar_bits, window_size, g2);
-        let g1_powers_projective = FixedBase::msm(scalar_bits, window_size, &g1_table, &tau_powers);
-        let g2_powers_projective = FixedBase::msm(scalar_bits, window_size, &g2_table, &tau_powers);
-        let g1_powers = P::G1::normalize_batch(&g1_powers_projective);
-        let g2_powers = P::G2::normalize_batch(&g2_powers_projective);
+        let (g1_powers_projective, g2_powers_projective) = rayon::join(
+            || {
+                let g1_table = FixedBase::get_window_table(scalar_bits, window_size, g1);
+                FixedBase::msm(scalar_bits, window_size, &g1_table, &beta_powers)
+            },
+            || {
+                let g2_table = FixedBase::get_window_table(scalar_bits, window_size, g2);
+                FixedBase::msm(scalar_bits, window_size, &g2_table, &beta_powers)
+            },
+        );
 
-        UniversalKzgSrs {
+        let (g1_powers, g2_powers) = rayon::join(
+            || P::G1::normalize_batch(&g1_powers_projective),
+            || P::G2::normalize_batch(&g2_powers_projective),
+        );
+
+        Self {
             g1_powers,
             g2_powers,
         }
     }
 
-    pub fn get_prover_key(&self, key_size: usize) -> Result<Vec<P::G1Affine>, KZGError> {
-        if self.g1_powers.len() < key_size {
-            return Err(KZGError::KeyLengthError(self.g1_powers.len(), key_size));
-        }
-        Ok(self.g1_powers[..=key_size].to_vec())
-    }
-
-    pub fn get_verifier_key(&self, key_size: usize) -> Result<KZGVerifierKey<P>, KZGError> {
-        if self.g1_powers.len() < key_size {
-            return Err(KZGError::KeyLengthError(self.g1_powers.len(), key_size));
-        }
-        Ok(KZGVerifierKey {
-            g1: self.g1_powers[0],
-            g2: self.g2_powers[0],
-            tau_2: self.g2_powers[1],
-        })
-    }
-
-    pub fn trim(&self, key_size: usize) -> Result<(Vec<P::G1Affine>, KZGVerifierKey<P>), KZGError> {
-        if self.g1_powers.len() < key_size {
-            return Err(KZGError::KeyLengthError(self.g1_powers.len(), key_size));
-        }
-        let g1_powers = self.g1_powers[..=key_size].to_vec();
-
-        let pk = g1_powers;
-        let vk = KZGVerifierKey {
-            g1: self.g1_powers[0],
-            g2: self.g2_powers[0],
-            tau_2: self.g2_powers[1],
-        };
-        Ok((pk, vk))
+    pub fn trim(params: Arc<Self>, max_degree: usize) -> (KZGProverKey<P>, KZGVerifierKey<P>) {
+        assert!(params.g1_powers.len() > 0, "max_degree is 0");
+        assert!(max_degree < params.g1_powers.len(), "SRS length is less than size");
+        let g1 = params.g1_powers[0];
+        let g2 = params.g2_powers[0];
+        let beta_g2 = params.g2_powers[1];
+        let pk = KZGProverKey::new(params, 0, max_degree + 1);
+        let vk = KZGVerifierKey { g1, g2, beta_g2 };
+        (pk, vk)
     }
 }
 
-pub struct UnivariateKZG<P> {
-    phantom: PhantomData<P>,
+#[derive(Clone, Debug)]
+pub struct KZGProverKey<P: Pairing> {
+    srs: Arc<SRS<P>>,
+    // offset to read into SRS
+    offset: usize,
+    // max size of srs
+    supported_size: usize,
 }
 
-impl<P: Pairing> UnivariateKZG<P>
+impl<P: Pairing> KZGProverKey<P> {
+    pub fn new(srs: Arc<SRS<P>>, offset: usize, supported_size: usize) -> Self {
+        assert!(
+            srs.g1_powers.len() >= offset + supported_size,
+            "not enough powers (req: {} from offset {}) in the SRS (length: {})",
+            supported_size,
+            offset,
+            srs.g1_powers.len()
+        );
+        Self {
+            srs,
+            offset,
+            supported_size,
+        }
+    }
+
+    pub fn g1_powers(&self) -> &[P::G1Affine] {
+        &self.srs.g1_powers[self.offset..self.offset + self.supported_size]
+    }
+}
+
+#[derive(Clone, Copy, Debug)]
+pub struct KZGVerifierKey<P: Pairing> {
+    pub g1: P::G1Affine,
+    pub g2: P::G2Affine,
+    pub beta_g2: P::G2Affine,
+}
+
+#[derive(Debug, Clone, Eq, PartialEq, Default)]
+pub struct UVKZGPCS<P: Pairing> {
+    _phantom: PhantomData<P>,
+}
+
+impl<P: Pairing> UVKZGPCS<P>
 where
-    <P as Pairing>::ScalarField: poly::field::JoltField,
+    <P as Pairing>::ScalarField: JoltField,
 {
     pub fn commit_offset(
-        g1_powers: &Vec<P::G1Affine>,
+        pk: &KZGProverKey<P>,
         poly: &UniPoly<P::ScalarField>,
         offset: usize,
-    ) -> Result<P::G1Affine, KZGError> {
-        if poly.degree() > g1_powers.len() {
-            return Err(KZGError::CommitLengthError(poly.degree(), g1_powers.len()));
+    ) -> Result<P::G1Affine, ProofVerifyError> {
+        if poly.degree() > pk.g1_powers().len() {
+            return Err(ProofVerifyError::KeyLengthError(
+                poly.degree(),
+                pk.g1_powers().len(),
+            ));
         }
 
-        let scalars = &poly.as_vec();
-        let bases = g1_powers.as_slice();
-
-        let com = <P::G1 as VariableBaseMSM>::msm(
+        let scalars = poly.as_vec();
+        let bases = pk.g1_powers();
+        let c = <P::G1 as VariableBaseMSM>::msm(
             &bases[offset..scalars.len()],
             &poly.as_vec()[offset..],
         )
         .unwrap();
 
-        Ok(com.into_affine())
+        Ok(c.into_affine())
     }
 
     pub fn commit(
-        g1_powers: &Vec<P::G1Affine>,
+        pk: &KZGProverKey<P>,
         poly: &UniPoly<P::ScalarField>,
-    ) -> Result<KZGCommitment<P>, KZGError> {
-        if poly.degree() > g1_powers.len() {
-            return Err(KZGError::CommitLengthError(poly.degree(), g1_powers.len()));
+    ) -> Result<P::G1Affine, ProofVerifyError> {
+        if poly.degree() > pk.g1_powers().len() {
+            return Err(ProofVerifyError::KeyLengthError(
+                poly.degree(),
+                pk.g1_powers().len(),
+            ));
         }
-        let com = <P::G1 as VariableBaseMSM>::msm(
-            &g1_powers.as_slice()[..poly.as_vec().len()],
-            &poly.as_vec(),
+        let c = <P::G1 as VariableBaseMSM>::msm(
+            &pk.g1_powers()[..poly.as_vec().len()],
+            &poly.as_vec().as_slice(),
         )
         .unwrap();
-        Ok(KZGCommitment(com.into_affine()))
+        Ok(c.into_affine())
     }
 
     pub fn open(
-        g1_powers: impl Borrow<Vec<P::G1Affine>>,
-        polynomial: &UniPoly<P::ScalarField>,
+        pk: &KZGProverKey<P>,
+        poly: &UniPoly<P::ScalarField>,
         point: &P::ScalarField,
-    ) -> Result<(P::G1Affine, P::ScalarField), KZGError> {
-        let g1_powers = g1_powers.borrow();
+    ) -> Result<(P::G1Affine, P::ScalarField), ProofVerifyError>
+    where
+        <P as ark_ec::pairing::Pairing>::ScalarField: JoltField,
+    {
         let divisor = UniPoly::from_coeff(vec![-*point, P::ScalarField::one()]);
-        let witness_polynomial = polynomial
-            .divide_with_q_and_r(&divisor)
-            .map(|(q, _r)| q)
-            .ok_or(KZGError::PolyDivisionError)?;
+        let (witness_poly, _) = poly.divide_with_q_and_r(&divisor).unwrap();
         let proof = <P::G1 as VariableBaseMSM>::msm(
-            &g1_powers.as_slice()[..witness_polynomial.as_vec().len()],
-            &witness_polynomial.as_vec(),
+            &pk.g1_powers()[..witness_poly.as_vec().len()],
+            &witness_poly.as_vec().as_slice(),
         )
         .unwrap();
-        let evaluation = polynomial.evaluate(point);
-
+        let evaluation = poly.evaluate(point);
         Ok((proof.into_affine(), evaluation))
-    }
-
-    fn _verify(
-        vk: impl Borrow<KZGVerifierKey<P>>,
-        commitment: &KZGCommitment<P>,
-        point: &P::ScalarField,
-        proof: &P::G1Affine,
-        evaluation: &P::ScalarField,
-    ) -> Result<bool, KZGError> {
-        let vk = vk.borrow();
-
-        let lhs = P::pairing(
-            commitment.0.into_group() - vk.g1.into_group() * evaluation,
-            vk.g2,
-        );
-        let rhs = P::pairing(proof, vk.tau_2.into_group() - (vk.g2 * point));
-        Ok(lhs == rhs)
     }
 }
 
 #[cfg(test)]
-mod tests {
+mod test {
     use super::*;
-    use crate::poly::unipoly::UniPoly;
     use ark_bn254::{Bn254, Fr};
+    use ark_ec::AffineRepr;
+    use ark_std::{rand::Rng, UniformRand};
+    use rand_chacha::ChaCha20Rng;
+    use rand_core::SeedableRng;
 
-    use ark_std::{
-        rand::{Rng, SeedableRng},
-        UniformRand,
-    };
-    use rand_chacha::{rand_core::RngCore, ChaCha20Rng};
-
-    fn random<P: Pairing, R: RngCore>(degree: usize, mut rng: &mut R) -> UniPoly<P::ScalarField>
-    where
-        <P as Pairing>::ScalarField: poly::field::JoltField,
-    {
-        let coeffs = (0..=degree)
-            .map(|_| P::ScalarField::rand(&mut rng))
-            .collect::<Vec<_>>();
-        UniPoly::from_coeff(coeffs)
+    fn kzg_verify<P: Pairing>(
+        vk: &KZGVerifierKey<P>,
+        commitment: &P::G1Affine,
+        point: &P::ScalarField,
+        proof: &P::G1Affine,
+        evaluation: &P::ScalarField,
+    ) -> Result<bool, ProofVerifyError> {
+        let lhs = P::pairing(
+            commitment.into_group() - vk.g1.into_group() * evaluation,
+            vk.g2,
+        );
+        let rhs = P::pairing(proof, vk.beta_g2.into_group() - (vk.g2 * point));
+        Ok(lhs == rhs)
     }
 
     #[test]
-    fn commit_prove_verify() -> Result<(), KZGError> {
+    fn kzg_commit_prove_verify() -> Result<(), ProofVerifyError> {
         let seed = b"11111111111111111111111111111111";
         for _ in 0..100 {
             let mut rng = &mut ChaCha20Rng::from_seed(*seed);
             let degree = rng.gen_range(2..20);
 
-            let pp = UniversalKzgSrs::<Bn254>::setup(degree, &mut rng);
-            let (ck, vk) = pp.trim(degree).unwrap();
-            let p = random::<Bn254, ChaCha20Rng>(degree, rng);
-            let comm = UnivariateKZG::<Bn254>::commit(&ck, &p)?;
+            let pp = Arc::new(SRS::<Bn254>::setup(&mut rng, degree));
+            let (ck, vk) = SRS::trim(pp, degree);
+            let p = UniPoly::random::<ChaCha20Rng>(degree, rng);
+            let comm = UVKZGPCS::<Bn254>::commit(&ck, &p)?;
             let point = Fr::rand(rng);
-            let (proof, value) = UnivariateKZG::<Bn254>::open(&ck, &p, &point)?;
+            let (proof, value) = UVKZGPCS::<Bn254>::open(&ck, &p, &point)?;
             assert!(
-                UnivariateKZG::<Bn254>::_verify(&vk, &comm, &point, &proof, &value)?,
+                kzg_verify(&vk, &comm, &point, &proof, &value)?,
                 "proof was incorrect for max_degree = {}, polynomial_degree = {}",
                 degree,
                 p.degree(),
