+++ jolt-core/src/poly/commitment/zeromorph.rs
@@ -4,18 +4,12 @@
 use std::{iter, marker::PhantomData};
 
 use crate::msm::VariableBaseMSM;
-use crate::poly::unipoly::UniPoly;
-use crate::poly::{self, dense_mlpoly::DensePolynomial};
-use crate::utils::errors::ProofVerifyError;
-use crate::utils::transcript::{AppendToTranscript, ProofTranscript};
-use ark_bn254::Bn254;
-use ark_ec::scalar_mul::fixed_base::FixedBase;
+use crate::poly::{self, unipoly::UniPoly, dense_mlpoly::DensePolynomial};
+use crate::utils::{errors::ProofVerifyError, transcript::{AppendToTranscript, ProofTranscript}};
 use ark_ec::{pairing::Pairing, AffineRepr, CurveGroup};
-use ark_ff::{batch_inversion, Field, PrimeField};
+use ark_ff::{batch_inversion, Field};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use ark_std::{One, UniformRand, Zero};
-use itertools::Itertools;
-use lazy_static::lazy_static;
+use ark_std::{One, Zero};
 use rand_chacha::{rand_core::SeedableRng, ChaCha20Rng};
 use rand_core::{CryptoRng, RngCore};
 use std::sync::Arc;
@@ -25,184 +19,7 @@ use ark_ec::VariableBaseMSM;
 
 use rayon::prelude::*;
 
-use super::commitment_scheme::{BatchType, CommitShape, CommitmentScheme};
-
-//use super::commitment_scheme::{ BatchType, CommitShape, CommitmentScheme};
-
-#[derive(Clone, Debug)]
-pub struct SRS<P: Pairing> {
-    pub g1_powers: Vec<P::G1Affine>,
-    pub g2_powers: Vec<P::G2Affine>,
-}
-
-impl<P: Pairing> SRS<P> {
-    pub fn setup<R: RngCore + CryptoRng>(mut rng: &mut R, max_degree: usize) -> Self {
-        let beta = P::ScalarField::rand(&mut rng);
-        let g1 = P::G1::rand(&mut rng);
-        let g2 = P::G2::rand(&mut rng);
-
-        let beta_powers: Vec<P::ScalarField> = (0..=max_degree)
-            .scan(beta, |acc, _| {
-                let val = *acc;
-                *acc *= beta;
-                Some(val)
-            })
-            .collect();
-
-        let window_size = FixedBase::get_mul_window_size(max_degree);
-        let scalar_bits = P::ScalarField::MODULUS_BIT_SIZE as usize;
-
-        //TODO: gate with rayon
-        let (g1_powers_projective, g2_powers_projective) = rayon::join(
-            || {
-                let g1_table = FixedBase::get_window_table(scalar_bits, window_size, g1);
-                FixedBase::msm(scalar_bits, window_size, &g1_table, &beta_powers)
-            },
-            || {
-                let g2_table = FixedBase::get_window_table(scalar_bits, window_size, g2);
-                FixedBase::msm(scalar_bits, window_size, &g2_table, &beta_powers)
-            },
-        );
-
-        let (g1_powers, g2_powers) = rayon::join(
-            || P::G1::normalize_batch(&g1_powers_projective),
-            || P::G2::normalize_batch(&g2_powers_projective),
-        );
-
-        Self {
-            g1_powers,
-            g2_powers,
-        }
-    }
-
-    pub fn trim(params: Arc<Self>, supported_size: usize) -> (KZGProverKey<P>, KZGVerifierKey<P>) {
-        assert!(params.g1_powers.len() > 0, "max_degree is 0");
-        let g1 = params.g1_powers[0];
-        let g2 = params.g2_powers[0];
-        let beta_g2 = params.g2_powers[1];
-        let pk = KZGProverKey::new(params, 0, supported_size + 1);
-        let vk = KZGVerifierKey { g1, g2, beta_g2 };
-        (pk, vk)
-    }
-}
-
-// Abstraction around SRS preventing copying. Arc of SRS
-#[derive(Clone, Debug)]
-pub struct KZGProverKey<P: Pairing> {
-    srs: Arc<SRS<P>>,
-    // offset to read into SRS
-    offset: usize,
-    // max size of srs
-    supported_size: usize,
-}
-
-impl<P: Pairing> KZGProverKey<P> {
-    pub fn new(srs: Arc<SRS<P>>, offset: usize, supported_size: usize) -> Self {
-        assert!(
-            srs.g1_powers.len() >= offset + supported_size,
-            "not enough powers (req: {} from offset {}) in the SRS (length: {})",
-            supported_size,
-            offset,
-            srs.g1_powers.len()
-        );
-        Self {
-            srs,
-            offset,
-            supported_size,
-        }
-    }
-
-    pub fn g1_powers(&self) -> &[P::G1Affine] {
-        &self.srs.g1_powers[self.offset..self.offset + self.supported_size]
-    }
-}
-
-// Abstraction around SRS preventing copying. Arc of SRS
-#[derive(Clone, Copy, Debug)]
-pub struct KZGVerifierKey<P: Pairing> {
-    pub g1: P::G1Affine,
-    pub g2: P::G2Affine,
-    pub beta_g2: P::G2Affine,
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Default)]
-pub struct UVKZGPCS<P: Pairing> {
-    _phantom: PhantomData<P>,
-}
-
-impl<P: Pairing> UVKZGPCS<P>
-where
-    <P as Pairing>::ScalarField: poly::field::JoltField,
-{
-    fn commit_offset(
-        pk: &KZGProverKey<P>,
-        poly: &UniPoly<P::ScalarField>,
-        offset: usize,
-    ) -> Result<P::G1Affine, ProofVerifyError> {
-        if poly.degree() > pk.g1_powers().len() {
-            return Err(ProofVerifyError::KeyLengthError(
-                poly.degree(),
-                pk.g1_powers().len(),
-            ));
-        }
-
-        let scalars = poly.as_vec();
-        let bases = pk.g1_powers();
-        let c = <P::G1 as VariableBaseMSM>::msm(
-            &bases[offset..scalars.len()],
-            &poly.as_vec()[offset..],
-        )
-        .unwrap();
-
-        Ok(c.into_affine())
-    }
-
-    pub fn commit(
-        pk: &KZGProverKey<P>,
-        poly: &UniPoly<P::ScalarField>,
-    ) -> Result<P::G1Affine, ProofVerifyError> {
-        if poly.degree() > pk.g1_powers().len() {
-            return Err(ProofVerifyError::KeyLengthError(
-                poly.degree(),
-                pk.g1_powers().len(),
-            ));
-        }
-        let c = <P::G1 as VariableBaseMSM>::msm(
-            &pk.g1_powers()[..poly.as_vec().len()],
-            &poly.as_vec().as_slice(),
-        )
-        .unwrap();
-        Ok(c.into_affine())
-    }
-
-    fn open(
-        pk: &KZGProverKey<P>,
-        poly: &UniPoly<P::ScalarField>,
-        point: &P::ScalarField,
-    ) -> Result<(P::G1Affine, P::ScalarField), ProofVerifyError>
-    where
-        <P as ark_ec::pairing::Pairing>::ScalarField: poly::field::JoltField,
-    {
-        let divisor = UniPoly::from_coeff(vec![-*point, P::ScalarField::one()]);
-        let (witness_poly, _) = poly.divide_with_q_and_r(&divisor).unwrap();
-        let proof = <P::G1 as VariableBaseMSM>::msm(
-            &pk.g1_powers()[..witness_poly.as_vec().len()],
-            &witness_poly.as_vec().as_slice(),
-        )
-        .unwrap();
-        let evaluation = poly.evaluate(point);
-        Ok((proof.into_affine(), evaluation))
-    }
-}
-
-const MAX_VARS: usize = 17;
-
-lazy_static! {
-    pub static ref ZEROMORPH_SRS: ZeromorphSRS<Bn254> = ZeromorphSRS(Arc::new(SRS::setup(
-        &mut ChaCha20Rng::from_seed(*b"ZEROMORPH_POLY_COMMITMENT_SCHEME"),
-        1 << (MAX_VARS + 1)
-    )));
-}
+use super::{commitment_scheme::{BatchType, CommitShape, CommitmentScheme}, kzg::{SRS, UVKZGPCS, KZGProverKey, KZGVerifierKey}};
 
 pub struct ZeromorphSRS<P: Pairing>(Arc<SRS<P>>);
 
@@ -212,7 +29,6 @@ impl<P: Pairing> ZeromorphSRS<P> {
     }
 
     pub fn trim(self, max_degree: usize) -> (ZeromorphProverKey<P>, ZeromorphVerifierKey<P>) {
-        //TODO: remove into()
         let (commit_pp, kzg_vk) = SRS::trim(self.0.clone(), max_degree);
         let offset = self.0.g1_powers.len() - max_degree;
         let tau_N_max_sub_2_N = self.0.g2_powers[offset];
@@ -435,7 +251,6 @@ where
         ))
     }
 
-    //TODO: change interface to create commitment for poly???
     pub fn open(
         pp: &ZeromorphProverKey<P>,
         poly: &DensePolynomial<P::ScalarField>,
@@ -453,7 +268,6 @@ where
             ));
         }
 
-        //assert_eq!(Self::commit(pp, poly).unwrap(), *comm);
         assert_eq!(poly.evaluate(point), *eval);
 
         let (quotients, remainder): (Vec<UniPoly<P::ScalarField>>, P::ScalarField) =
@@ -462,16 +276,15 @@ where
         assert_eq!(remainder, *eval);
 
         // Compute the multilinear quotients q_k = q_k(X_0, ..., X_{k-1})
+        // TODO: multicore gate
         let q_k_com: Vec<P::G1Affine> = quotients
             .par_iter()
             .map(|q| UVKZGPCS::commit(&pp.commit_pp, q).unwrap())
             .collect();
         let q_comms: Vec<P::G1> = q_k_com
-            .clone()
-            .into_iter()
+            .par_iter()
             .map(|c| c.into_group())
             .collect();
-        //transcript.append_points(b"q_comms", &q_comms);
         q_comms
             .iter()
             .for_each(|c| transcript.append_point(b"quo", c));
@@ -521,10 +334,6 @@ where
         })
     }
 
-    //Batch together polynomials -> Then commit
-    // polys[0..m]
-    // commitments[0..m]
-    // evals[0..m]
     fn batch_open(
         pk: &ZeromorphProverKey<P>,
         polynomials: &[&DensePolynomial<P::ScalarField>],
@@ -534,6 +343,8 @@ where
     ) -> ZeromorphProof<P> {
         let num_vars = point.len();
         let n = 1 << num_vars;
+
+        //TODO(pat): produce powers in parallel 
         // Generate batching challenge \rho and powers 1,...,\rho^{m-1}
         let rho: P::ScalarField = transcript.challenge_scalar(b"rho");
         // Compute batching of unshifted polynomials f_i, and batched eval v_i:
@@ -550,8 +361,8 @@ where
                 (f_batched, batched_evaluation)
             },
         );
-        let pi_poly = DensePolynomial::new(f_batched.Z.clone());
-        Zeromorph::<P>::open(&pk, &pi_poly, &point, &batched_evaluation, transcript).unwrap()
+        let poly = DensePolynomial::new(f_batched.Z.clone());
+        Zeromorph::<P>::open(&pk, &poly, &point, &batched_evaluation, transcript).unwrap()
     }
 
     fn batch_verify(
@@ -562,6 +373,8 @@ where
         batch_proof: &ZeromorphProof<P>,
         transcript: &mut ProofTranscript,
     ) -> Result<(), ProofVerifyError> {
+
+        //TODO(pat): produce powers in parallel using window method
         // Compute batching of unshifted polynomials f_i:
         // Compute powers of batching challenge rho
         let rho: P::ScalarField = transcript.challenge_scalar(b"rho");
@@ -585,7 +398,6 @@ where
         )
     }
 
-    //Change api
     pub fn verify(
         vk: &ZeromorphVerifierKey<P>,
         comm: &ZeromorphCommitment<P>,
@@ -596,12 +408,9 @@ where
     ) -> Result<(), ProofVerifyError> {
         transcript.append_protocol_name(Self::protocol_name());
 
-        // Receive commitments [q_k]
-        //TODO: remove clone
         let q_comms: Vec<P::G1> = proof
             .q_k_com
-            .clone()
-            .into_iter()
+            .iter()
             .map(|c| c.into_group())
             .collect();
         q_comms
@@ -636,7 +445,6 @@ where
         .concat();
         let bases = [
             vec![proof.q_hat_com, comm.0, vk.kzg_vk.g1],
-            //TODO: eliminate
             proof.q_k_com.clone(),
         ]
         .concat();
@@ -669,11 +477,13 @@ where
     type Proof = ZeromorphProof<P>;
     type BatchedProof = ZeromorphProof<P>;
 
-    fn setup(_shapes: &[CommitShape]) -> Self::Setup {
+    fn setup(shapes: &[CommitShape]) -> Self::Setup {
+        let max_len = shapes.iter().map(|shape| shape.input_length).max().unwrap();
+
         ZeromorphSRS(Arc::new(SRS::setup(
             &mut ChaCha20Rng::from_seed(*b"ZEROMORPH_POLY_COMMITMENT_SCHEME"),
-            65536 + 1
-        ))).trim(65536)
+            max_len
+        ))).trim(max_len)
     }
 
     fn commit(poly: &DensePolynomial<Self::Field>, setup: &Self::Setup) -> Self::Commitment {
@@ -696,8 +506,8 @@ where
         let iter = evals.par_iter();
         #[cfg(not(feature = "multicore"))]
         let iter = evals.iter();
-        iter.enumerate()
-            .map(|(i, evals)| {
+        iter
+            .map(|evals| {
                 assert!(
                     gens.0.commit_pp.g1_powers().len() > evals.len(),
                     "COMMIT KEY LENGTH ERROR {}, {}", gens.0.commit_pp.g1_powers().len(), evals.len()
@@ -722,7 +532,6 @@ where
         opening_point: &[Self::Field], // point at which the polynomial is evaluated
         transcript: &mut ProofTranscript,
     ) -> Self::Proof {
-        //TODO: setup
         let eval = poly.evaluate(&opening_point);
         Zeromorph::<P>::open(
             &setup.0,
@@ -739,10 +548,9 @@ where
         polynomials: &[&DensePolynomial<Self::Field>],
         opening_point: &[Self::Field],
         openings: &[Self::Field],
-        batch_type: BatchType,
+        _batch_type: BatchType,
         transcript: &mut ProofTranscript,
     ) -> Self::BatchedProof {
-        //TODO: setup
         Zeromorph::<P>::batch_open(
             &setup.0,
             polynomials,
@@ -798,9 +606,8 @@ mod test {
     use super::*;
     use crate::utils::math::Math;
     use ark_bn254::{Bn254, Fr};
-    use ark_ec::AffineRepr;
     use ark_ff::{BigInt, Zero};
-    use ark_std::{rand::Rng, test_rng, UniformRand};
+    use ark_std::{test_rng, UniformRand};
     use rand_core::SeedableRng;
 
     // Evaluate Phi_k(x) = \sum_{i=0}^k x^i using the direct inefficent formula
@@ -1050,54 +857,6 @@ mod test {
         }
     }
 
-    fn kzg_verify<P: Pairing>(
-        vk: &KZGVerifierKey<P>,
-        commitment: &P::G1Affine,
-        point: &P::ScalarField,
-        proof: &P::G1Affine,
-        evaluation: &P::ScalarField,
-    ) -> Result<bool, ProofVerifyError> {
-        let lhs = P::pairing(
-            commitment.into_group() - vk.g1.into_group() * evaluation,
-            vk.g2,
-        );
-        let rhs = P::pairing(proof, vk.beta_g2.into_group() - (vk.g2 * point));
-        Ok(lhs == rhs)
-    }
-
-    fn random<P: Pairing, R: RngCore>(degree: usize, mut rng: &mut R) -> UniPoly<P::ScalarField>
-    where
-        <P as Pairing>::ScalarField: poly::field::JoltField,
-    {
-        let coeffs = (0..=degree)
-            .map(|_| P::ScalarField::rand(&mut rng))
-            .collect::<Vec<_>>();
-        UniPoly::from_coeff(coeffs)
-    }
-
-    #[test]
-    fn kzg_commit_prove_verify() -> Result<(), ProofVerifyError> {
-        let seed = b"11111111111111111111111111111111";
-        for _ in 0..100 {
-            let mut rng = &mut ChaCha20Rng::from_seed(*seed);
-            let degree = rng.gen_range(2..20);
-
-            let pp = Arc::new(SRS::<Bn254>::setup(&mut rng, degree));
-            let (ck, vk) = SRS::trim(pp, degree);
-            let p = random::<Bn254, ChaCha20Rng>(degree, rng);
-            let comm = UVKZGPCS::<Bn254>::commit(&ck, &p)?;
-            let point = Fr::rand(rng);
-            let (proof, value) = UVKZGPCS::<Bn254>::open(&ck, &p, &point)?;
-            assert!(
-                kzg_verify(&vk, &comm, &point, &proof, &value)?,
-                "proof was incorrect for max_degree = {}, polynomial_degree = {}",
-                degree,
-                p.degree(),
-            );
-        }
-        Ok(())
-    }
-
     #[test]
     fn zeromorph_commit_prove_verify() {
         for num_vars in [4, 5, 6] {
@@ -1167,11 +926,11 @@ mod test {
 
                 let srs = ZeromorphSRS::<Bn254>::setup(&mut rng, 1 << num_vars);
                 let (pk, vk) = srs.trim(1 << num_vars);
-                let commitments: Vec<ZeromorphCommitment<Bn254>> = polys
+                let commitments: Vec<_> = polys
                     .iter()
                     .map(|poly| Zeromorph::<Bn254>::commit(&pk, &poly).unwrap())
                     .collect();
-
+ 
                 let commitments_refs: Vec<_> = commitments.iter().map(|x| x).collect();
                 let polys_refs: Vec<_> = polys.iter().map(|x| x).collect();
